erro pendente
