+++ pokedex/src/App.js
@@ -1,13 +1,14 @@
 import { MainContainer } from "./styled"
 import React from "react";
 import Router from "./Router/Router"
+import GlobalState from './global/GlobalState'
 
 const App = () => {
 
   return (
-    <MainContainer>
+    <GlobalState>
       <Router />
-    </MainContainer>
+    </GlobalState>
 
   )
 }
