+++ pokedex/src/pages/HomePage/HomePage.js
@@ -1,11 +1,19 @@
-import React from "react";
+import React, { useEffect } from "react";
 import { useHistory } from "react-router-dom";
 import { goToPokedex } from "../../Router/coordinator";
-import Card from '../../Components/Cards'
+import Card from '../../components/Cards'
+import { GlobalStateContext } from '../global/GlobalStateContext'
+import { useContext } from 'react'
+
 
 
 const HomePage = () => {
+    const { states, setters, requests } = useContext(GlobalStateContext)
     const history = useHistory();
+
+    useEffect(() => {
+        requests.getAllPokemons();
+    }, [requests])
 
     return (
         <div>
