+++ pokedex/src/pages/Pokedex/Pokedex.js
@@ -1,7 +1,9 @@
-import Card from "../../Components/Cards";
-import { Cards } from "./styled";
+import Card from "../../components/Cards";
+// import { GlobalStateContext } from '../global/GlobalState'
+// import { useContext } from 'react'
 
 const Pokedex = () => {
+    // const { states, setters, requests } = useContext(GlobalStateContext)
 
     return (
         <div>
